requirements.txt
flask
pymysql
sqlalchemy
gunicorn